# [Enhancement] Avoid repeated calculation of bitmap memory usage