--- a/be/src/exprs/bitmap_functions.cpp
+++ b/be/src/exprs/bitmap_functions.cpp
@@ -569,7 +569,7 @@ StatusOr<ColumnPtr> BitmapFunctions::bitmap_to_base64(FunctionContext* context, 
 
     for (int row = 0; row < size; ++row) {
         BitmapValue* bitmap = viewer.value(row);
-        int byteSize = bitmap->getSizeInBytes();
+        int byteSize = bitmap->get_size_in_bytes();
         std::unique_ptr<char[]> buf;
         buf.reset(new char[byteSize]);
 
@@ -684,7 +684,7 @@ StatusOr<ColumnPtr> BitmapFunctions::bitmap_to_binary(FunctionContext* context, 
     raw::RawString buf;
     for (int row = 0; row < size; ++row) {
         BitmapValue* bitmap = viewer.value(row);
-        size_t serialize_size = bitmap->getSizeInBytes();
+        size_t serialize_size = bitmap->get_size_in_bytes();
         buf.resize(serialize_size);
         bitmap->write(buf.data());
         builder.append(Slice(buf.data(), serialize_size));