--- a/be/src/types/bitmap_value.cpp
+++ b/be/src/types/bitmap_value.cpp
@@ -1095,7 +1095,7 @@ int64_t BitmapValue::bitmap_subset_limit_internal(const int64_t& range_start, co
             auto end = _bitmap->begin();
 
             int64_t offset = 0;
-            for (; end != _bitmap->end() && offset < abs_limit && * end <= range_start;) {
+            for (; end != _bitmap->end() && offset < abs_limit && *end <= range_start;) {
                 ++end;
                 ++offset;
             }