--- a/be/src/types/hll.h
+++ b/be/src/types/hll.h
@@ -141,6 +141,8 @@ public:
 
     uint64_t serialize_size() const { return max_serialized_size(); }
 
+    uint64_t mem_usage() const { return max_serialized_size(); }
+
     // common interface
     void clear();