--- a/be/src/util/percentile_value.h
+++ b/be/src/util/percentile_value.h
@@ -41,6 +41,8 @@ public:
         return 1 + _tdigest.serialize_size();
     }
 
+    uint64_t mem_usage() const { return 1 + _tdigest.serialize_size(); }
+
     size_t serialize(uint8_t* writer) const {
         *(writer) = _type;
         return _tdigest.serialize(writer + 1);