--- a/be/test/exprs/bitmap_functions_test.cpp
+++ b/be/test/exprs/bitmap_functions_test.cpp
@@ -48,7 +48,7 @@ TEST_F(VecBitmapFunctionsTest, bitmapEmptyTest) {
 
         auto* bitmap = ColumnHelper::get_const_value<TYPE_OBJECT>(column);
 
-        ASSERT_EQ(1, bitmap->getSizeInBytes());
+        ASSERT_EQ(1, bitmap->get_size_in_bytes());
     }
 }