--- a/be/test/types/bitmap_value_test.cpp
+++ b/be/test/types/bitmap_value_test.cpp
@@ -66,6 +66,7 @@ void BitmapValueTest::check_bitmap(BitmapDataType type, const BitmapValue& bitma
     for (auto i = start; i < end; i++) {
         ASSERT_TRUE(bitmap.contains(i));
     }
+    ASSERT_EQ(bitmap.mem_usage(), bitmap.serialize_size());
 }
 
 void BitmapValueTest::check_bitmap(BitmapDataType type, const BitmapValue& bitmap, uint64_t start_1, uint64_t end_1,
@@ -78,6 +79,7 @@ void BitmapValueTest::check_bitmap(BitmapDataType type, const BitmapValue& bitma
     for (auto i = start_2; i < end_2; i++) {
         ASSERT_TRUE(bitmap.contains(i));
     }
+    ASSERT_EQ(bitmap.mem_usage(), bitmap.serialize_size());
 }
 
 TEST_F(BitmapValueTest, copy_construct) {
@@ -96,6 +98,7 @@ TEST_F(BitmapValueTest, assign_operator) {
     bitmap_1.add(64);
     check_bitmap(BitmapDataType::BITMAP, bitmap_1, 0, 65);
     check_bitmap(BitmapDataType::BITMAP, _large_bitmap, 0, 64);
+    ASSERT_EQ(bitmap_1.mem_usage(), bitmap_1.serialize_size());
 
     BitmapValue bitmap_2 = _medium_bitmap;
     bitmap_2.add(14);
@@ -572,7 +575,7 @@ TEST_F(BitmapValueTest, bitmap_min) {
 
 TEST_F(BitmapValueTest, bitmap_serialize_deserialize) {
     // empty bitmap
-    size_t size = _empty_bitmap.getSizeInBytes();
+    size_t size = _empty_bitmap.get_size_in_bytes();
     char buf_1[size];
     _empty_bitmap.write(buf_1);
     BitmapValue bitmap_1;
@@ -581,7 +584,7 @@ TEST_F(BitmapValueTest, bitmap_serialize_deserialize) {
     check_bitmap(BitmapDataType::EMPTY, bitmap_1, 0, 0);
 
     // single bitmap
-    size = _single_bitmap.getSizeInBytes();
+    size = _single_bitmap.get_size_in_bytes();
     char buf_2[size];
     _single_bitmap.write(buf_2);
     BitmapValue bitmap_2;
@@ -590,7 +593,7 @@ TEST_F(BitmapValueTest, bitmap_serialize_deserialize) {
     check_bitmap(BitmapDataType::SINGLE, bitmap_2, 0, 1);
 
     // medium bitmap
-    size = _medium_bitmap.getSizeInBytes();
+    size = _medium_bitmap.get_size_in_bytes();
     char buf_3[size];
     _medium_bitmap.write(buf_3);
     BitmapValue bitmap_3;
@@ -599,7 +602,7 @@ TEST_F(BitmapValueTest, bitmap_serialize_deserialize) {
     check_bitmap(BitmapDataType::SET, bitmap_3, 0, 14);
 
     // large bitmap
-    size = _large_bitmap.getSizeInBytes();
+    size = _large_bitmap.get_size_in_bytes();
     char buf_4[size];
     _large_bitmap.write(buf_4);
     BitmapValue bitmap_4;
@@ -610,7 +613,7 @@ TEST_F(BitmapValueTest, bitmap_serialize_deserialize) {
 
 TEST_F(BitmapValueTest, test_valid_and_deserialize) {
     // empty bitmap
-    size_t size = _empty_bitmap.getSizeInBytes();
+    size_t size = _empty_bitmap.get_size_in_bytes();
     char buf_1[size];
     _empty_bitmap.write(buf_1);
     BitmapValue bitmap_1;
@@ -619,7 +622,7 @@ TEST_F(BitmapValueTest, test_valid_and_deserialize) {
     check_bitmap(BitmapDataType::EMPTY, bitmap_1, 0, 0);
 
     // single bitmap
-    size = _single_bitmap.getSizeInBytes();
+    size = _single_bitmap.get_size_in_bytes();
     char buf_2[size];
     _single_bitmap.write(buf_2);
     BitmapValue bitmap_2;
@@ -628,7 +631,7 @@ TEST_F(BitmapValueTest, test_valid_and_deserialize) {
     check_bitmap(BitmapDataType::SINGLE, bitmap_2, 0, 1);
 
     // medium bitmap
-    size = _medium_bitmap.getSizeInBytes();
+    size = _medium_bitmap.get_size_in_bytes();
     char buf_3[size];
     _medium_bitmap.write(buf_3);
     BitmapValue bitmap_3;
@@ -637,7 +640,7 @@ TEST_F(BitmapValueTest, test_valid_and_deserialize) {
     check_bitmap(BitmapDataType::SET, bitmap_3, 0, 14);
 
     // large bitmap
-    size = _large_bitmap.getSizeInBytes();
+    size = _large_bitmap.get_size_in_bytes();
     char buf_4[size];
     _large_bitmap.write(buf_4);
     BitmapValue bitmap_4(_large_bitmap);
@@ -1011,7 +1014,7 @@ TEST_F(BitmapValueTest, next_batch) {
 
 std::string convert_bitmap_to_string(BitmapValue& bitmap) {
     std::string buf;
-    buf.resize(bitmap.getSizeInBytes());
+    buf.resize(bitmap.get_size_in_bytes());
     bitmap.write((char*)buf.c_str());
     return buf;
 }