## Why I'm doing:

Currently, `BitmapValue` already supports copy on write, but `Column::mem_usage` will still repeatedly calculate mem_usage of same bitmap. The cost of bitmap calculating mem_usage is relatively high, so currently a field mem_usage is added to `BitmapValue` to prevent repeated calculation of mem_usage.

performance test:

create table and prepare data

```
mysql> select count(*) from lineorder;
+-----------+
| count(*)  |
+-----------+
| 143999468 |
+-----------+

CREATE TABLE `bitmap_1m` (
  `k1` int(11) NULL COMMENT "",
  `v1` bitmap BITMAP_UNION NULL COMMENT ""
) ENGINE=OLAP 
AGGREGATE KEY(`k1`)
DISTRIBUTED BY HASH(`k1`) BUCKETS 1 
PROPERTIES (
"replication_num" = "1",
"in_memory" = "false",
"enable_persistent_index" = "false",
"replicated_storage" = "true",
"fast_schema_evolution" = "true",
"compression" = "LZ4"
); 

insert into bitmap_1m select lo_partkey%500 as c1, bitmap_agg(lo_orderkey) from lineorder group by c1;
```

query sql:

```
select count(lo_partkey), count(v1) from lineorder join bitmap_1m on lo_linenumber=k1 and bitmap_contains(v1, lo_orderkey);
```

before opt:

```
TotalTime: 19.328s
ChunkAccumulateTime: 5.714s
```

after opt:

```
TotalTime: 13.584s
ChunkAccumulateTime: 38.451ms
```

## What I'm doing:

1. Add item _mem_usage to `BitmapValue` to prevent repeat mem usage calculation.
2. Rename `getSizeInBytes` to `get_size_in_bytes`.
3. Add interface mem_usage for `BitmapValue`, `HyperLogLog`, `JsonValue`, `PercentileValue`.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.2
  - [x] 3.1
  - [x] 3.0
  - [x] 2.5